moved mixin to a separate file
